+++ lib/project_types/extension/features/argo_renderer_package.rb
@@ -12,17 +12,13 @@ module Extension
         ARGO_ADMIN,
         ARGO_POST_PURCHASE,
       ].freeze
-      MINIMUM_ARGO_VERSION = "0.9.3".freeze
 
       property! :package_name, accepts: PACKAGE_NAMES
       property! :version, accepts: String
 
       class << self
-        def from_package_manager(package_manager_output)
-          pattern = /(?<name>#{PACKAGE_NAMES.join("|")})@(?<version>\d.*)$/
-          match = package_manager_output.match(pattern)
-          raise PackageNotFound, package_manager_output if match.nil?
-          new(package_name: match[:name], version: match[:version].strip)
+        def from_npm_package(package)
+          new(package_name: package.name, version: package.version)
         end
       end
 
@@ -32,15 +28,6 @@ module Extension
 
       def admin?
         package_name == ARGO_ADMIN
-      end
-
-      ##
-      # Temporarily returns false in all cases as the argo webpack server is
-      # unable to handle the UUID flag.
-      def supports_uuid_flag?
-        false
-        # return false if checkout?
-        # Gem::Version.new(version) > Gem::Version.new(MINIMUM_ARGO_VERSION)
       end
     end
   end