+++ test/shopify-cli/method_object_test.rb
@@ -36,6 +36,22 @@ module ShopifyCli
       end
     end
 
+    class BuildArray
+      include ShopifyCli::MethodObject
+
+      def call(*elements)
+        elements
+      end
+    end
+
+    class TransformArray
+      include ShopifyCli::MethodObject
+
+      def call(*elements, &transform)
+        elements.map(&transform)
+      end
+    end
+
     def test_returns_a_result
       GenerateHelloWorld.new.call.tap do |result|
         assert_kind_of(ShopifyCli::Result::Success, result)
@@ -79,6 +95,24 @@ module ShopifyCli
             assert_equal "John", person.firstname
             assert_equal "Doe", person.lastname
           end
+        end
+    end
+
+    def test_does_not_forward_an_empty_list_of_keword_arguments_to_call
+      BuildArray
+        .call("Hello", "World")
+        .tap do |result|
+          assert_predicate(result, :success?)
+          assert_equal ["Hello", "World"], result.value
+        end
+    end
+
+    def test_forwards_blocks_to_call
+      TransformArray
+        .call("Hello", "World", &:upcase)
+        .tap do |result|
+          assert_predicate(result, :success?)
+          assert_equal %w[HELLO WORLD], result.value
         end
     end
   end